Adapted QCWF scripts to match PBS_QUEUE on convey* rather than "convey", to ensure operation oth new cluster with convey_fast, convey_medium and convey_long queues.
Summary:
Changed all test on whether script runs on convey to match on extended PBS_QUEUE=convey*. Centralized this query at the beginning of the scripts.

Centralized the ON_CONVEY variable value into a function runningOnConvey() in the bashLib.sh.

Test Plan:
As usual the original and modified versions of the workflow (convey, noconvey; 1.0.178, 1.0.180, current[these modifications]) were run on a test-file (WGS) and the outcomes was compared based on MD5 sums and inspection. No differences were found due to the
changes in these commits.

Reviewers: buchhalt, heinold

Reviewed By: heinold

Differential Revision: http://tbi-cb14.inet.dkfz-heidelberg.de/D90